int & long optimized + int overflow example

+++ CodeForces/Round642.cs
@@ -91,11 +91,11 @@ namespace CodeForces
             for (int j = 0; j < t; j++)
             {
                 s = Console.ReadLine().Trim();
-                long n = long.Parse(s);
+                int n = int.Parse(s);
                 long sum = 0;
-                for (long i = n - 1; i > 0; i = i - 2)
+                for (int i = n - 1; i > 0; i = i - 2)
                 {
-                    sum = sum + i * 2 * i;
+                    sum = sum + (long)i * 2 * i;
                 }
 
                 Console.WriteLine(sum);
